The scheduler now accepts shifts as a paremeter and listen to its click events
As the title said, the library now accepts an array of shifts as a input and display them on the scheduler. Also the library offers support for dayClick and also dayEvent click. So if we want to capture the info of the clicked event, now we can. By the way, if we click over the day, we will obtain all the shifts contained by that day.